// frontend/src/api/model/coreType.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Apicius API
 * OpenAPI spec version: 1.0.0
 */

export type CoreType = typeof CoreType[keyof typeof CoreType];


export const CoreType = {
  TEXT: 'TEXT',
  WHOLE_NUMBER: 'WHOLE_NUMBER',
  DECIMAL_NUMBER: 'DECIMAL_NUMBER',
  YES_NO: 'YES_NO',
  DATE: 'DATE',
  DATE_TIME: 'DATE_TIME',
} as const;