// frontend/src/api/model/refinement.ts
/**
 * Generated by orval v8.19.0 🍺
 * Do not edit manually.
 * Apicius API
 * OpenAPI spec version: 1.0.0
 */

export type Refinement = typeof Refinement[keyof typeof Refinement];


export const Refinement = {
  EMAIL: 'EMAIL',
  UUID: 'UUID',
  URL: 'URL',
  PASSWORD: 'PASSWORD',
  INT32: 'INT32',
  INT64: 'INT64',
  FLOAT: 'FLOAT',
  DOUBLE: 'DOUBLE',
} as const;